+++ docs/obsidian/vault/.obsidian/workspace.json
@@ -149,10 +149,10 @@
   },
   "active": "f770321e399e1be0",
   "lastOpenFiles": [
+    "Learned Things/MySQL.md",
     "Learned Things/npm.md",
     "Learned Things/Project.md",
     "Welcome.md",
-    "Learned Things/MySQL.md",
     "Learned Things/images/Pasted image 20240721224716.png",
     "Learned Things/images",
     "Learned Things.md",
