+++ server/src/endpoints/Endpoints.ts
@@ -1,0 +1,5 @@
+import * as POST from './POST/POST' 
+export {POST}
+
+import * as GET from './GET/GET'
+export {GET}
