+++ server/src/endpoints/GET/GET.ts
@@ -1,0 +1,1 @@
+export {userBudget} from './UserBudget'
