+++ server/src/endpoints/POST/POST.ts
@@ -1,0 +1,1 @@
+export {loginUser} from './Login'
